+++ src/js/components/auths/signUp/template.js
