+++ src/js/components/auths/signUp/validate.js
