+++ src/js/components/auths/signUp/view.js
