+++ src/js/components/contents/lines/view.js
