+++ src/js/components/home/template.js
