+++ src/js/components/home/view.js
